Poprawki cd

+++ Silownia/Silownia/Controllers/SalaController.cs
@@ -19,17 +19,17 @@ namespace Silownia.Controllers
         // GET: /Sala/
         public ActionResult Index()
         {
-         //   if(Session["User"] != null)
+            if(Session["Auth"].ToString() == "Recepcjonista" | Session["Auth"].ToString() == "Administrator")
             {
                 return View(db.Sale.ToList());
             }
-          //  return HttpNotFound();
+            return HttpNotFound();
         }
 
          // GET: /Sala/Details/5
         public ActionResult Details(int? id)
         {
-         //   if (Session["User"] != null)
+            if (Session["Auth"].ToString() == "Recepcjonista" | Session["Auth"].ToString() == "Administrator")
             {
                 if (id == null)
                 {
@@ -42,18 +42,18 @@ namespace Silownia.Controllers
                 }
                 return View(sala);
             }
-          //  return HttpNotFound();
+            return HttpNotFound();
         }
 
         // GET: /Sala/Create
         public ActionResult Create()
         {
-          //  if (Session["User"] != null)
+            if (Session["Auth"].ToString() == "Recepcjonista" | Session["Auth"].ToString() == "Administrator")
             {
                 ViewBag.SilowniaID = new SelectList(db.Silownie, "SilowniaID", "Nazwa");
                 return View();
             }
-         //   return HttpNotFound();
+            return HttpNotFound();
         }
 
         // POST: /Sala/Create
@@ -63,7 +63,7 @@ namespace Silownia.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create(Sala sala, HttpPostedFileBase file)
         {
-         //   if (Session["User"] != null)
+            if (Session["Auth"].ToString() == "Recepcjonista" | Session["Auth"].ToString() == "Administrator")
             {
                 if (ModelState.IsValid)
                 {
@@ -102,7 +102,7 @@ namespace Silownia.Controllers
                 ViewBag.SilowniaID = new SelectList(db.Silownie, "SilowniaID", "Nazwa");
                 return View(sala);
             }
-         //   return HttpNotFound();
+            return HttpNotFound();
         }
 
         public ActionResult Zdjecie(int id, int img)
@@ -117,7 +117,7 @@ namespace Silownia.Controllers
         // GET: /Sala/Edit/5
         public ActionResult Edit(int? id)
         {
-         //   if (Session["User"] != null)
+            if (Session["Auth"].ToString() == "Recepcjonista" | Session["Auth"].ToString() == "Administrator")
             {
                 if (id == null)
                 {
@@ -131,7 +131,7 @@ namespace Silownia.Controllers
                 ViewBag.SilowniaID = new SelectList(db.Silownie, "SilowniaID", "Nazwa", sala.Silownia);
                 return View(sala);
             }
-         //   return HttpNotFound();
+            return HttpNotFound();
         }
 
         // POST: /Sala/Edit/5
@@ -141,7 +141,7 @@ namespace Silownia.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include="Numer_sali,Rodzaj,Status,Opis,ImageFile,SilowniaID")] Sala sala)
         {
-         //   if (Session["User"] != null)
+            if (Session["Auth"].ToString() == "Recepcjonista" | Session["Auth"].ToString() == "Administrator")
             {
                 if (ModelState.IsValid)
                 {
@@ -152,13 +152,13 @@ namespace Silownia.Controllers
                 ViewBag.SilowniaID = new SelectList(db.Silownie, "SilowniaID", "Nazwa", sala.Silownia);
                 return View(sala);
             }
-          //  return HttpNotFound();
+            return HttpNotFound();
         }
 
         // GET: /Sala/Delete/5
         public ActionResult Delete(int? id)
         {
-         //   if (Session["User"] != null)
+            if (Session["Auth"].ToString() == "Recepcjonista" | Session["Auth"].ToString() == "Administrator")
             {
                 if (id == null)
                 {
@@ -171,7 +171,7 @@ namespace Silownia.Controllers
                 }
                 return View(sala);
             }
-          //  return HttpNotFound();
+            return HttpNotFound();
         }
 
         // POST: /Sala/Delete/5
@@ -179,14 +179,14 @@ namespace Silownia.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-         //   if (Session["User"] != null)
+            if (Session["Auth"].ToString() == "Recepcjonista" | Session["Auth"].ToString() == "Administrator")
             {
                 Sala sala = db.Sale.Find(id);
                 db.Sale.Remove(sala);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-         //   return HttpNotFound();
+         /   return HttpNotFound();
         }
 
         protected override void Dispose(bool disposing)
